Redirect PIE802 to C419
