+++ crates/ruff/resources/test/fixtures/flake8_comprehensions/C419.py
@@ -1,4 +1,3 @@
-# PIE802
 any([x.id for x in bar])
 all([x.id for x in bar])
 any(  # first comment
@@ -14,6 +13,7 @@ all(x.id for x in bar)
 all(x.id for x in bar)
 any(x.id for x in bar)
 all((x.id for x in bar))
+
 
 async def f() -> bool:
     return all([await use_greeting(greeting) for greeting in await greetings()])
