+++ crates/ruff/src/registry.rs
@@ -228,23 +228,24 @@ ruff_macros::register_rules!(
     // flake8-blind-except
     rules::flake8_blind_except::rules::BlindExcept,
     // flake8-comprehensions
+    rules::flake8_comprehensions::rules::UnnecessaryCallAroundSorted,
+    rules::flake8_comprehensions::rules::UnnecessaryCollectionCall,
+    rules::flake8_comprehensions::rules::UnnecessaryComprehension,
+    rules::flake8_comprehensions::rules::UnnecessaryComprehensionAnyAll,
+    rules::flake8_comprehensions::rules::UnnecessaryDoubleCastOrProcess,
+    rules::flake8_comprehensions::rules::UnnecessaryGeneratorDict,
     rules::flake8_comprehensions::rules::UnnecessaryGeneratorList,
     rules::flake8_comprehensions::rules::UnnecessaryGeneratorSet,
-    rules::flake8_comprehensions::rules::UnnecessaryGeneratorDict,
-    rules::flake8_comprehensions::rules::UnnecessaryListComprehensionSet,
-    rules::flake8_comprehensions::rules::UnnecessaryListComprehensionDict,
-    rules::flake8_comprehensions::rules::UnnecessaryLiteralSet,
-    rules::flake8_comprehensions::rules::UnnecessaryLiteralDict,
-    rules::flake8_comprehensions::rules::UnnecessaryCollectionCall,
-    rules::flake8_comprehensions::rules::UnnecessaryLiteralWithinTupleCall,
-    rules::flake8_comprehensions::rules::UnnecessaryLiteralWithinListCall,
     rules::flake8_comprehensions::rules::UnnecessaryListCall,
-    rules::flake8_comprehensions::rules::UnnecessaryCallAroundSorted,
-    rules::flake8_comprehensions::rules::UnnecessaryDoubleCastOrProcess,
-    rules::flake8_comprehensions::rules::UnnecessarySubscriptReversal,
-    rules::flake8_comprehensions::rules::UnnecessaryComprehension,
-    rules::flake8_comprehensions::rules::UnnecessaryMap,
+    rules::flake8_comprehensions::rules::UnnecessaryListComprehensionDict,
+    rules::flake8_comprehensions::rules::UnnecessaryListComprehensionSet,
+    rules::flake8_comprehensions::rules::UnnecessaryLiteralDict,
+    rules::flake8_comprehensions::rules::UnnecessaryLiteralSet,
     rules::flake8_comprehensions::rules::UnnecessaryLiteralWithinDictCall,
+    rules::flake8_comprehensions::rules::UnnecessaryLiteralWithinListCall,
+    rules::flake8_comprehensions::rules::UnnecessaryLiteralWithinTupleCall,
+    rules::flake8_comprehensions::rules::UnnecessaryMap,
+    rules::flake8_comprehensions::rules::UnnecessarySubscriptReversal,
     // flake8-debugger
     rules::flake8_debugger::rules::Debugger,
     // mccabe
@@ -570,7 +571,6 @@ ruff_macros::register_rules!(
     rules::flake8_pie::rules::UnnecessaryDictKwargs,
     rules::flake8_pie::rules::ReimplementedListBuiltin,
     rules::flake8_pie::rules::MultipleStartsEndsWith,
-    rules::flake8_pie::rules::UnnecessaryComprehensionAnyAll,
     // flake8-commas
     rules::flake8_commas::rules::MissingTrailingComma,
     rules::flake8_commas::rules::TrailingCommaOnBareTuple,
