+++ crates/ruff/src/rule_redirects.rs
@@ -92,5 +92,6 @@ static REDIRECTS: Lazy<HashMap<&'static str, &'static str>> = Lazy::new(|| {
         ("TYP001", "TCH001"),
         // TODO(charlie): Remove by 2023-06-01.
         ("RUF004", "B026"),
+        ("PIE802", "C419"),
     ])
 });
