+++ crates/ruff/src/rules/flake8_pie/mod.rs
@@ -1,5 +1,4 @@
 //! Rules from [flake8-pie](https://pypi.org/project/flake8-pie/).
-mod fixes;
 pub(crate) mod rules;
 
 #[cfg(test)]
@@ -21,7 +20,6 @@ mod tests {
     #[test_case(Rule::UnnecessarySpread, Path::new("PIE800.py"); "PIE800")]
     #[test_case(Rule::ReimplementedListBuiltin, Path::new("PIE807.py"); "PIE807")]
     #[test_case(Rule::NonUniqueEnums, Path::new("PIE796.py"); "PIE796")]
-    #[test_case(Rule::UnnecessaryComprehensionAnyAll, Path::new("PIE802.py"); "PIE802")]
     fn rules(rule_code: Rule, path: &Path) -> Result<()> {
         let snapshot = format!("{}_{}", rule_code.noqa_code(), path.to_string_lossy());
         let diagnostics = test_path(
