+++ ruff.schema.json
@@ -1605,6 +1605,7 @@
         "C416",
         "C417",
         "C418",
+        "C419",
         "C9",
         "C90",
         "C901",
@@ -1920,7 +1921,6 @@
         "PIE8",
         "PIE80",
         "PIE800",
-        "PIE802",
         "PIE804",
         "PIE807",
         "PIE81",
